pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "langchain>=0.0.27",
    "discord.py>=2.3.1",
    "aiosupabase>=0.0.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true



[tool.pyright]
exclude = [ "./.venv" ]
venvPath = "."
venv = "./.venv"
